% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyFMDV.R
\name{ClassifyInfinity}
\alias{ClassifyInfinity}
\title{Classify FMDV pipeline}
\usage{
ClassifyInfinity(
  inputFile,
  outputFile = "results.csv",
  model,
  QC_value = 0.6,
  Length_value = 0.5,
  N_value = 2,
  write_result = FALSE
)
}
\arguments{
\item{inputFile}{string path relative to the working directory of the input file. Must be in fasta format.}

\item{outputFile}{string path relative to the working directory of the output file. Default = "results.csv"}

\item{model}{A random forest classification model. Must be FMDV_model. See \link{FMDV_model}.}

\item{QC_value}{numeric value from 0 to 1. (default = 0.4)}

\item{Length_value}{numeric value from 0 to 1. Proportion of difference to the expected sequence length. (default = 0.5)}

\item{N_value}{numeric value from 0 to 100. Percentage of acceptable ambiguous bases. (default = 2)}

\item{write_result}{Logical. If True writes result as a .csv file. If false returns a data.frame.}
}
\value{
a comma-delimited file or a data.frame
}
\description{
Wrapper function that performs all the recommended step. This includes
\describe{
\item{Kcounter}{# Count Kmers and normalize data}
\item{PredictionCaller}{Run prediction pipeline}
\item{QualityControl}{Check Quality of the input data and classification results}
\item{Quality_filter}{For sequences with any FLAG present set result Clade to LowQuality}
}
}
\examples{

\dontrun{
file_path<-system.file("extdata","KY404934.1.fasta",package="voRtex")

ClassifyInfinity(inputFile=file_path,model=FMDV_model)
}
}
