% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coverage.R
\name{compute_coverage}
\alias{compute_coverage}
\title{compute_coverage
returns data containing the average position and coverage}
\usage{
compute_coverage(inputdata, windowsize = 100, logarize = TRUE)
}
\arguments{
\item{inputdata}{dataframe}

\item{windowsize}{calculates the average coverage of contiguous windows of size windowsize}

\item{logarize}{takes the base 10 log of coverage values}
}
\value{
returns data containing the average position and coverage
}
\description{
compute_coverage
returns data containing the average position and coverage
}
\examples{
 FilePath <- system.file("extdata", "SRR12664421_full_coverage.bed",
                         package = "vortex", mustWork = TRUE)
data <- read.table(FilePath, col.names = c("reference", "startpos", "endpos", "coverage"))
data_processed<-compute_coverage(data, 50,TRUE)
data_processed
}
