% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenbank.R
\name{create_gb_list}
\alias{create_gb_list}
\title{create_gb_list}
\usage{
create_gb_list(x)
}
\arguments{
\item{x}{DNAstringset with the relevant sequence}
}
\value{
A list
}
\description{
Creates a simple list containing all the necessary information for the \code{\link[=writeGenBank]{writeGenBank()}} function.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "vortex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)
create_gb_list(sequence_stringset)
}
