% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameBinder.R
\name{DataFrameBinder}
\alias{DataFrameBinder}
\title{DataFrameBinder}
\usage{
DataFrameBinder(vcflist)
}
\arguments{
\item{vcflist}{A list of collapsedVCF objects to combine in a single data.frame using
\code{\link[=VCFToDataFrame]{VCFToDataFrame()}}}
}
\value{
A \code{\link[=data.frame]{data.frame()}}
}
\description{
DataFrameBinder
}
\examples{
file <- system.file("extdata", "variant_file.vcf", package = "voRtex", mustWork = TRUE)
file2 <- system.file("extdata", "variant_file_2.vcf", package = "voRtex", mustWork = TRUE)
vcf_list <- list(
  file,
  file2
)
listavcf_salida <- list()
for (i in 1:length(vcf_list)) {
  listavcf_salida[[i]] <- VariantAnnotation::readVcf(vcf_list[[i]])
}
DataFrameBinder(listavcf_salida)

}
