% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenbank.R
\name{find_ORF}
\alias{find_ORF}
\title{find_ORF
Finds the ATG initiation codon in FMDV genomes.}
\usage{
find_ORF(x, start_codon = "ATG")
}
\arguments{
\item{x}{DNAstringset with the relevant sequence}

\item{start_codon}{Pattern string to search. Default ATG.}
}
\value{
the position of the relevant codon
}
\description{
find_ORF
Finds the ATG initiation codon in FMDV genomes.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "voRtex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)
find_ORF(sequence_stringset)

}
