% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenBank.R
\name{find_STOP}
\alias{find_STOP}
\title{find_STOP
Finds the first stop codon in FMDV genomes. It uses start_codon_position
to determine the correct translation frame.}
\usage{
find_STOP(x, start_codon_position)
}
\arguments{
\item{x}{DNAstringset with the relevant sequence}

\item{start_codon_position}{numeric. Output from \code{\link[=find_ORF]{find_ORF()}}.}
}
\value{
the position of the relevant codon
}
\description{
find_STOP
Finds the first stop codon in FMDV genomes. It uses start_codon_position
to determine the correct translation frame.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "voRtex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)
find_STOP(sequence_stringset,
          start_codon_position=find_ORF(sequence_stringset))
}
