% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.feature.by.position.R
\name{get.feature.by.position}
\alias{get.feature.by.position}
\title{get.feature.by.position}
\usage{
get.feature.by.position(features, query)
}
\arguments{
\item{features}{a data.frame with at least three columns: Feature, Start and End. Describes the features of the genome and their coordinates.
Can be created with \link{import.features.to.df}}

\item{query}{a numeric that indicates the nucleotide position.}
}
\value{
A string with the feature name at the query position
}
\description{
get.feature.by.position
}
\examples{
features_fmdv <- import.features.to.df()
get.feature.by.position(features_fmdv,2000)
}
