% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenbank.R
\name{get_polyprotein}
\alias{get_polyprotein}
\title{get_polyprotein}
\usage{
get_polyprotein(x, start_codon = NULL, stop_codon = NULL)
}
\arguments{
\item{x}{DNAstringset with the relevant sequence}

\item{start_codon}{numeric. Position of initiation codon.}

\item{stop_codon}{numeric. Last position of stop codon.}
}
\value{
AA sequence
}
\description{
A very simple wrapper around \link[Biostrings:translate]{Biostrings::translate}. Translate a DNA sequence
If no initiatio and/or stop codons positions are provided it uses \code{\link[=find_ORF]{find_ORF()}} and \code{\link[=find_STOP]{find_STOP()}}.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "vortex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)
get_polyprotein(sequence_stringset)

}
