% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_heatmap.R
\name{ggplot_heatmap}
\alias{ggplot_heatmap}
\title{ggplot_heatmap
create a custom heatmap plot using ggplot2.}
\usage{
ggplot_heatmap(
  inputdata,
  color_pal,
  name = "Sample",
  low_limit = 0,
  high_limit = 5
)
}
\arguments{
\item{inputdata}{data used to create the heatmap chart.}

\item{color_pal}{vector specifying the colors used in the color scale of the heat map}

\item{name}{name of the sample seen in the graph}

\item{low_limit}{lower and upper limits of the color scale}

\item{high_limit}{lower and upper limits of the color scale}
}
\value{
heatmap chart
}
\description{
ggplot_heatmap
create a custom heatmap plot using ggplot2.
}
\examples{
data <- read.table("Resultados/run1/coverage/SRR13776172_full_coverage.bed",
        col.names = c("reference", "startpos", "endpos", "coverage"))
data_processed<-compute_coverage(data, 50, TRUE)

ggplot_heatmap(inputdata=data_processed,
              color_pal = color)
}
