% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.feature.by.position.R
\name{import.features.to.df}
\alias{import.features.to.df}
\title{import.features.to.df}
\usage{
import.features.to.df(
  file = system.file("extdata", "fmdv_features.csv", package = "voRtex")
)
}
\arguments{
\item{file}{path to a .csv file to import the genome features. Must include the columns: Feature, Start and End.}
}
\value{
a dataframe with the relevant features
}
\description{
import.features.to.df
}
\examples{
features_fmdv <- import.features.to.df()
}
