% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lector.runs.R
\name{lector.runs}
\alias{lector.runs}
\title{lector.runs
lee archivos csv de Sra, y crea archivos de texto en
paquetes especificados por el usuario, con
la informacion de los runs}
\usage{
lector.runs(
  input.file = NULL,
  folder = "Salida",
  output.file = "prueba.txt",
  chunklength = 100
)
}
\arguments{
\item{input.file}{string con la direccion relativa del archivo}

\item{folder}{string con el nombre de la carpeta de salida (deafult=salida)}

\item{output.file}{string con el nombre del archivo de salida (default=prueba.txt)}

\item{chunklength}{numeric con maximo numero de muestras por archivo (default=100)}
}
\value{
multiples archivos de texto con la informacion de la columna run
}
\description{
lector.runs
lee archivos csv de Sra, y crea archivos de texto en
paquetes especificados por el usuario, con
la informacion de los runs
}
\examples{
 lector.runs("vortex/datos_crudos2/SraRunInfo_completa_12_12_22.csv",
 output.file = "run",
 chunklength = 250)
}
