% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ModelControl}
\alias{ModelControl}
\title{ModelControl
Utility function to check the classification model object}
\usage{
ModelControl(model)
}
\arguments{
\item{model}{random forest classification model. Must be FMDV_model. See \link{FMDV_model}.}
}
\value{
warnings
}
\description{
ModelControl
Utility function to check the classification model object
}
\examples{
ModelControl(FMDV_model)
}
