% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_from_string.R
\name{name_from_string}
\alias{name_from_string}
\title{name_from_string
extracts the relevant part of a string delimited by first_token and last_token}
\usage{
name_from_string(
  sample_string,
  first_token = "extdata/",
  last_token = "_masked.fasta",
  verbose = FALSE
)
}
\arguments{
\item{sample_string}{String. Sample string to extract from.-}

\item{first_token}{String.first part to cut}

\item{last_token}{String.last part to cut}

\item{verbose}{Boolean. Sets verbose mode (default: false)}
}
\value{
String. Relevant part of the string.
}
\description{
name_from_string
extracts the relevant part of a string delimited by first_token and last_token
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "vortex", mustWork = TRUE)
name_from_string(file,
  first_token = "extdata/",
  last_token = "_masked.fasta"
)

}
