% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCountFilter.R
\name{NCountFilter}
\alias{NCountFilter}
\title{NCountFilter
Filters objects from the class DNAStringSet, using as parameter the number
of "N" bases}
\usage{
NCountFilter(dna_seq, filter = 1000)
}
\arguments{
\item{dna_seq}{Object from the class DNAStringSet}

\item{filter}{Numeric to use as a filter}
}
\value{
Object from the class DNAStringSet
}
\description{
NCountFilter
Filters objects from the class DNAStringSet, using as parameter the number
of "N" bases
}
\examples{
FilePath <- system.file("extdata",
                        "renamed_all.fasta",
                        package = "voRtex",
                        mustWork = TRUE)

DNASequence <- Biostrings::readDNAStringSet(FilePath)

NCountFilter(DNASequence)

}
