% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NreadFilter.R
\name{NreadFilter}
\alias{NreadFilter}
\title{NreadFilter
reads a dataframe created from a bed file and filters it according to the
number of reads.}
\usage{
NreadFilter(OGDataFrame, RFilter = 1000)
}
\arguments{
\item{OGDataFrame}{DataFrame.A dataframe Created from a bed file,
that must have the columns "reference", "startpos", "endpos"
and "nreads"}

\item{RFilter}{Numeric. Minimum read number}
}
\value{
List. With 2 objects: a dataframe, the og dataframe filtered by
number of reads, and with a new column named "width", and a numeric
that says the percentage that passed the filter
}
\description{
NreadFilter
reads a dataframe created from a bed file and filters it according to the
number of reads.
}
\examples{
FilePath <- system.file("extdata", "SRR12664421_full_coverage.bed",
                         package = "voRtex", mustWork = TRUE)
OGDataFrame <- read.table(FilePath,
                          col.names = c("reference","startpos","endpos","nreads"))
salida <- NreadFilter(OGDataFrame,5000)
}
