% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectControl.R
\name{objectControl}
\alias{objectControl}
\title{objectControl
Controls if the correct class is indicated}
\usage{
objectControl(object)
}
\arguments{
\item{object}{class object to control}
}
\value{
null
}
\description{
objectControl
Controls if the correct class is indicated
}
