% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{Position}
\alias{Position}
\title{Position
Returns a vector with the allele positions}
\usage{
Position(vcf_data)
}
\arguments{
\item{vcf_data}{CollapsedVCF vcf file}
}
\value{
Vector with the allele positions
}
\description{
Position
Returns a vector with the allele positions
}
\examples{
file <- system.file("extdata", "variant_file.vcf", package = "voRtex", mustWork = TRUE)
vcf_data <- VariantAnnotation::readVcf(file)
Position(vcf_data)
}
