% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{QualityControl}
\alias{QualityControl}
\title{QualityControl}
\usage{
QualityControl(data, QC_value = 0.4, Length_value = 0.5, N_value = 2, model)
}
\arguments{
\item{data}{data.frame obtained with \code{\link[=PredictionCaller]{PredictionCaller()}}}

\item{QC_value}{numeric value from 0 to 1. (default = 0.4)}

\item{Length_value}{numeric value from 0 to 1. Proportion of difference to the expected sequence length. (default = 0.5)}

\item{N_value}{numeric value from 0 to 100. Percentage of acceptable ambiguous bases. (default = 2)}

\item{model}{A random forest classification model. Must be FMDV_model. See \link{FMDV_model}.}
}
\value{
A list with three logical vectors. In each case TRUE means pass.
}
\description{
QualityControl
}
\examples{

file_path<-system.file("extdata","KY404934.1.fasta",package="voRtex")

sequence<-ape::read.FASTA(file_path,type = "DNA")

calling_null<-ranger::predictions(FMDV_model)

NormalizedData <- Kcounter(SequenceData=sequence,model=FMDV_model)

PredictedData <- PredictionCaller(NormalizedData=NormalizedData,model=FMDV_model)

QualityControl(PredictedData,model=FMDV_model)


}
