% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_vcf.R
\name{recalibrate_vcf}
\alias{recalibrate_vcf}
\title{recalibrate_vcf
Take a vcf object and update the position of each SNP based on the coordintates of a
aligned sequence. Each gap in the sequence adds a nt to the relevant position.}
\usage{
recalibrate_vcf(vcf_data, DNA_stringset)
}
\arguments{
\item{vcf_data}{collapsedVCF object. Loaded with the \link[VariantAnnotation:readVcf-methods]{VariantAnnotation::readVcf}.}

\item{DNA_stringset}{DNA_stringset. Loaded with \link[Biostrings:XStringSet-io]{Biostrings::readDNAStringSet}.}
}
\value{
an updated collapsedVCF object.
}
\description{
recalibrate_vcf
Take a vcf object and update the position of each SNP based on the coordintates of a
aligned sequence. Each gap in the sequence adds a nt to the relevant position.
}
\examples{
FilePath <- system.file("extdata", "aligned_sample.fasta",
                         package = "voRtex", mustWork = TRUE)
DNA_stringset<-Biostrings::readDNAStringSet(FilePath)
vcf_path<-  system.file("extdata", "vcf_to_update.vcf",
                         package = "voRtex", mustWork = TRUE)
Vcf_object<-VariantAnnotation::readVcf(vcf_path)
recalibrate_vcf(Vcf_object,
               DNA_stringset)

}
