% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run voRtex shiny app}
\usage{
runShinyApp()
}
\value{
a shiny app
}
\description{
Deploys a server that runs the voRtex app locally
}
\examples{
\dontrun{
voRtex::runShinyApp()
}

}
