% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fasta_header.R
\name{update_fasta_header}
\alias{update_fasta_header}
\title{update_fasta_header
updates the Fasta header by a user-given string.}
\usage{
update_fasta_header(DNA_stringset, header)
}
\arguments{
\item{DNA_stringset}{DNA_stringset object. The fasta file to update}

\item{header}{String. The new header (name).}
}
\value{
DNA_stringset.
}
\description{
update_fasta_header
updates the Fasta header by a user-given string.
}
\examples{

file <- system.file("extdata", "SRR12664421_masked.fasta", package = "voRtex", mustWork = TRUE)
DNA_stringset <- Biostrings::readDNAStringSet(file)
header <- name_from_string(file)
update_fasta_header(
  DNA_stringset,
  header
)

}
