% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortex-package.R
\docType{package}
\name{vortex-package}
\alias{vortex}
\alias{vortex-package}
\title{vortex: voRtex}
\description{
A set of functions for processing VCF files in the context of datamining from illumina data.
}
\author{
\strong{Maintainer}: Marco Cacciabue \email{marcocacciabue@yahoo.com} (\href{http://orcid.org/0000-0002-1429-4252}{ORCID})

Authors:
\itemize{
  \item Melina Obregon \email{meliobre06@gmail.com} (\href{http://orcid.org/0000-0002-1429-4252}{ORCID})
  \item Nahuel Fenoglio \email{pequefenoglio99@gmail.com} (\href{http://orcid.org/0000-0002-1429-4252}{ORCID})
}

}
\keyword{internal}
