% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_renamed_fasta.R
\name{write_updated_fasta}
\alias{write_updated_fasta}
\title{write_updated_fasta
wrraper around \code{\link[=update_fasta_header]{update_fasta_header()}} reads a fasta file from path, updates the header and writes the new file}
\usage{
write_updated_fasta(file, output_dir, first_token, last_token)
}
\arguments{
\item{file}{string. The path of the fasta file.}

\item{output_dir}{string. Where to save the new fasta.}

\item{first_token}{String.first part to cut}

\item{last_token}{String.last part to cut}
}
\description{
write_updated_fasta
wrraper around \code{\link[=update_fasta_header]{update_fasta_header()}} reads a fasta file from path, updates the header and writes the new file
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "voRtex", mustWork = TRUE)
\dontrun{
write_updated_fasta(file,
output_dir="out",
first_token = "extdata/",
last_token = "_masked.fasta")
}

}
