% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenbank.R
\name{writeFeatureTable}
\alias{writeFeatureTable}
\title{writeFeatureTable
Additional function to create Feature table. Useful for using table2asn.exe software, in case \code{\link[=writeGenBank]{writeGenBank()}} doesn´t work.}
\usage{
writeFeatureTable(x, file = "out.tbl")
}
\arguments{
\item{x}{a list created with \code{\link[=create_gb_list]{create_gb_list()}}.}

\item{file}{string. Name of the file to save. Default = out.tbl.}
}
\value{
writes a feature table file
}
\description{
writeFeatureTable
Additional function to create Feature table. Useful for using table2asn.exe software, in case \code{\link[=writeGenBank]{writeGenBank()}} doesn´t work.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "voRtex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)

\dontrun{
writeFeatureTable(create_gb_list(sequence_stringset))
}
}
