% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenBank.R
\name{writeGenBank}
\alias{writeGenBank}
\title{writeGenBank
Creates a bare bones gb format file. It is intended to be used for the SNPeff database creation
step. Do NOT use it to upload to ncbi.}
\usage{
writeGenBank(x, file = "out.gb")
}
\arguments{
\item{x}{A list created with \code{\link[=create_gb_list]{create_gb_list()}}. It has all the relevant information
to add to the genbank file.}

\item{file}{string. Name of the file to save. Default = out.gb.}
}
\value{
writes a genbank-like file
}
\description{
writeGenBank
Creates a bare bones gb format file. It is intended to be used for the SNPeff database creation
step. Do NOT use it to upload to ncbi.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "voRtex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)
x<-create_gb_list(sequence_stringset)
file_out<-paste0(x$name,".gb")
\dontrun{
writeGenBank(x,file_out)
}
}
