% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenBank.R
\name{writeSequence}
\alias{writeSequence}
\title{writeSequence
write fasta sequence in genbank format.}
\usage{
writeSequence(x, file = "out.gb")
}
\arguments{
\item{x}{a list created with \code{\link[=create_gb_list]{create_gb_list()}}.}

\item{file}{string. Name of the file to save. Default = out.gb.}
}
\value{
writes the sequence section of the genbank file
}
\description{
writeSequence
write fasta sequence in genbank format.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "voRtex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)

\dontrun{
writeSequence(create_gb_list(sequence_stringset))
}
}
