% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGenbank.R
\name{writeSequenceAA}
\alias{writeSequenceAA}
\title{writeSequenceAA
write AA sequence in the genbank format, under the translation slot.}
\usage{
writeSequenceAA(x, file = "out.gb")
}
\arguments{
\item{x}{a list created with \code{\link[=create_gb_list]{create_gb_list()}}.}

\item{file}{string. Name of the file to save. Default = out.gb.}
}
\value{
writes the translation section of the genbank file
}
\description{
writeSequenceAA
write AA sequence in the genbank format, under the translation slot.
}
\examples{
file <- system.file("extdata", "SRR12664421_masked.fasta", package = "vortex", mustWork = TRUE)
sequence_stringset <- Biostrings::readDNAStringSet(file)
 \dontrun{
writeSequenceAA(create_gb_list(sequence_stringset))
}
}
