% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_element_from_ANN.R
\name{extract_element_from_ANN}
\alias{extract_element_from_ANN}
\title{extract_element_from_ANN}
\usage{
extract_element_from_ANN(vcf_data, index)
}
\arguments{
\item{vcf_data}{collapsedVCF object. Loaded with the \code{\link[VariantAnnotation:readVcf-methods]{VariantAnnotation::readVcf()}}.}

\item{index}{A numeric that indicates the index of the element of ANN to extract (tested for 2 and 3).}
}
\value{
A character vector
}
\description{
extract_element_from_ANN
}
\examples{
file <- system.file("extdata", "annotated.vcf", package = "voRtex", mustWork = TRUE)

vcf<- VariantAnnotation::readVcf(file)

Annotation<-extract_element_from_ANN(vcf_data=vcf, index=2)
Annotation_Impact<-extract_element_from_ANN(vcf_data=vcf,index=3)
}
