% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_heatmap.R
\name{ggplot_heatmap}
\alias{ggplot_heatmap}
\title{ggplot_heatmap
create a custom heatmap plot using ggplot2.}
\usage{
ggplot_heatmap(
  inputdata,
  color_pal,
  name = "Sample",
  low_limit = 0,
  high_limit = 5
)
}
\arguments{
\item{inputdata}{data used to create the heatmap chart.}

\item{color_pal}{vector specifying the colors used in the color scale of the heat map}

\item{name}{name of the sample seen in the graph}

\item{low_limit}{lower and upper limits of the color scale}

\item{high_limit}{lower and upper limits of the color scale}
}
\value{
heatmap chart
}
\description{
ggplot_heatmap
create a custom heatmap plot using ggplot2.
}
\examples{

 FilePath <- system.file("extdata", "SRR12664421_full_coverage.bed",
                         package = "vortex", mustWork = TRUE)
data <- read.table(FilePath, col.names = c("reference", "startpos", "endpos", "coverage"))
data_processed<-compute_coverage(data, 50, TRUE)
color  <- c("#D53E4F","#F46D43","#FDAE61","#FEE08B","#E6F598","#ABDDA4","#66C2A5","#3288BD")
ggplot_heatmap(inputdata=data_processed,
              color_pal = color)
}
