% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameBinder.R
\name{read.vcf.to.df}
\alias{read.vcf.to.df}
\title{read.vcf.to.df}
\usage{
read.vcf.to.df(file)
}
\arguments{
\item{file}{A string path to the vcf file to read. Created to reemplace \link{DataFrameBinder}.}
}
\value{
A data.frame created with \link{VCFToDataFrame}.
}
\description{
read.vcf.to.df
}
\examples{
file <- system.file("extdata", "variant_file.vcf", package = "voRtex", mustWork = TRUE)
read.vcf.to.df(file)

}
