% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameBinder.R
\name{read.vcf.to.df}
\alias{read.vcf.to.df}
\title{read.vcf.to.df}
\usage{
read.vcf.to.df(file, add_sample = FALSE, sample_name)
}
\arguments{
\item{file}{A string path to the vcf file to read. Created to reemplace \link{DataFrameBinder}.}

\item{add_sample}{Logical. if true adds \link{sample_name} in a new column.}

\item{sample_name}{string. Name of sample to add if \link{add_sample} is TRUE.}
}
\value{
A \link{data.frame} created with \link{VCFToDataFrame}.
}
\description{
read.vcf.to.df
}
\examples{
file <- system.file("extdata", "variant_file.vcf", package = "voRtex", mustWork = TRUE)
read.vcf.to.df(file)

}
