% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFToDataFrame.R
\name{VCFToDataFrame}
\alias{VCFToDataFrame}
\title{VCFToDataFrame
Returns a data frame with the Positions, DP, AF and ANN extracted from a file.vcf}
\usage{
VCFToDataFrame(vcf_data)
}
\arguments{
\item{vcf_data}{collapsedVCF object. Loaded with the \link[VariantAnnotation:readVcf-methods]{VariantAnnotation::readVcf}.}
}
\value{
\link{Data.frame}
}
\description{
VCFToDataFrame
Returns a data frame with the Positions, DP, AF and ANN extracted from a file.vcf
}
\examples{
file <- system.file("extdata", "variant_file.vcf", package = "voRtex", mustWork = TRUE)
vcf_data <- VariantAnnotation::readVcf(file)
VCFToDataFrame(vcf_data)

}
